refactor: extract WelcomePage widget
